@@ -1,13 +1,13 @@
 {
   "name": "@dontcode/plugin-common",
-  "version": "0.5.15-34",
+  "version": "0.6.1",
   "peerDependencies": {
-    "@angular/common": "^13.3.2",
-    "@angular/core": "^13.3.2",
-    "@dontcode/core": "^0.5.1",
-    "rxjs": "^7.5.5"
+    "@angular/common": "^15.1.1",
+    "@angular/core": "^15.1.1",
+    "@dontcode/core": "^0.6.1",
+    "rxjs": "^7.8.0"
   },
   "dependencies": {
-    "tslib": "^2.3.1"
+    "tslib": "^2.4.1"
   }
 }
