@@ -2,13 +2,16 @@
   "name": "@dontcode/plugin-fields",
   "version": "0.5.15-29",
   "peerDependencies": {
-    "@angular/common": "^13.3.2",
-    "@angular/core": "^13.3.2",
-    "@dontcode/core": "^0.5.1",
-    "@dontcode/plugin-common": "^0.5.0",
-    "rxjs": "^7.5.5"
+    "@angular/common": "^15.1.1",
+    "@angular/core": "^15.1.1",
+    "@dontcode/core": "^0.6.1",
+    "@dontcode/plugin-common": "^0.6.1",
+    "all-country-data": "^1.0.3",
+    "flag-icon-css": "^4.1.7",
+    "i18n-iso-countries": "^7.5.0",
+    "rxjs": "^7.8.0"
   },
   "dependencies": {
-    "tslib": "^2.3.1"
+    "tslib": "^2.4.1"
   }
 }
