@@ -4,85 +4,87 @@
   "license": "MIT",
   "scripts": {
     "ng": "nx",
-    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2015 browser module main",
+    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2020 browser module main",
     "start": "nx serve",
     "build": "nx build",
     "test": "nx test",
+    "build-all": "nx run common:build:development && nx run sandbox:build:development && nx run basic:build:development && nx run fields:build:development",
     "push-all": "cd libs/common && npm version prerelease && cd ../sandbox && npm version prerelease && cd ../basic && npm version prerelease && cd ../fields && npm version prerelease && cd ../.. && nx run common:build:development && nx run sandbox:build:development && nx run basic:build:development && nx run fields:build:development && cd dist/libs/common && yalc push && cd ../sandbox && yalc push && cd ../basic && yalc push && cd ../fields && yalc push",
     "push": "cd libs/${npm_config_dir} && npm version prerelease && cd ../.. && nx run ${npm_config_dir}:build:development && cd dist/libs/${npm_config_dir} && yalc push"
   },
   "private": true,
   "dependencies": {
-    "@angular-architects/module-federation-runtime": "~14.2.0",
-    "@angular/animations": "^13.3.2",
-    "@angular/cdk": "^13.3.2",
-    "@angular/common": "^13.3.2",
-    "@angular/compiler": "^13.3.2",
-    "@angular/core": "^13.3.2",
-    "@angular/forms": "^13.3.2",
-    "@angular/platform-browser": "^13.3.2",
-    "@angular/platform-browser-dynamic": "^13.3.2",
-    "@angular/router": "^13.3.2",
+    "@angular-architects/module-federation-runtime": "^15.0.3",
+    "@angular/animations": "^15.1.1",
+    "@angular/cdk": "^15.1.1",
+    "@angular/common": "^15.1.1",
+    "@angular/compiler": "^15.1.1",
+    "@angular/core": "^15.1.1",
+    "@angular/forms": "^15.1.1",
+    "@angular/platform-browser": "^15.1.1",
+    "@angular/platform-browser-dynamic": "^15.1.1",
+    "@angular/router": "^15.1.1",
     "@dontcode/core": "file:.yalc/@dontcode/core",
-    "@fortawesome/angular-fontawesome": "^0.10.2",
-    "@fortawesome/fontawesome-svg-core": "^6.1.1",
-    "@fortawesome/free-regular-svg-icons": "^6.1.1",
-    "@fortawesome/free-solid-svg-icons": "^6.1.1",
-    "@ngrx/component-store": "13.1.0",
-    "all-country-data": "^0.2.2",
-    "async-mutex": "^0.3.2",
-    "broadcast-channel": "^4.10.0",
-    "core-js": "^3.21.1",
+    "@fortawesome/angular-fontawesome": "^0.12.1",
+    "@fortawesome/fontawesome-svg-core": "^6.2.1",
+    "@fortawesome/free-regular-svg-icons": "^6.2.1",
+    "@fortawesome/free-solid-svg-icons": "^6.2.1",
+    "@ngrx/component-store": "^15.1.0",
+    "all-country-data": "^1.0.3",
+    "async-mutex": "^0.4.0",
+    "broadcast-channel": "^4.20.2",
+    "core-js": "^3.27.2",
     "dexie": "^3.2.2",
     "flag-icon-css": "^4.1.7",
-    "i18n-iso-countries": "^7.4.0",
-    "jsonpath-plus": "^7.0.0",
-    "primeflex": "^3.2.1",
-    "primeicons": "^5.0.0",
-    "primeng": "^13.4.1",
-    "rxjs": "^7.5.5",
-    "tslib": "^2.3.1",
-    "zone.js": "0.11.5"
+    "i18n-iso-countries": "^7.5.0",
+    "jsonpath-plus": "^7.2.0",
+    "primeflex": "^3.3.0",
+    "primeicons": "^6.0.1",
+    "primeng": "^15.1.1",
+    "rxjs": "^7.8.0",
+    "tslib": "^2.4.1",
+    "zone.js": "^0.12.0"
   },
   "devDependencies": {
-    "@angular-architects/module-federation": "~14.2.0",
-    "@angular-devkit/build-angular": "^13.3.2",
-    "@angular-eslint/eslint-plugin": "^13.2.0",
-    "@angular-eslint/eslint-plugin-template": "^13.2.0",
-    "@angular-eslint/template-parser": "^13.2.0",
-    "@angular/cli": "^13.3.2",
-    "@angular/compiler-cli": "^13.3.2",
-    "@angular/language-service": "^13.3.2",
-    "@nrwl/angular": "14.0.3",
-    "@nrwl/cli": "14.0.3",
-    "@nrwl/cypress": "14.0.3",
-    "@nrwl/eslint-plugin-nx": "14.0.3",
-    "@nrwl/jest": "14.0.3",
-    "@nrwl/linter": "14.0.3",
-    "@nrwl/workspace": "14.0.3",
-    "@types/jest": "27.4.1",
-    "@types/node": "^16.11.7",
-    "@typescript-eslint/eslint-plugin": "^5.18.0",
-    "@typescript-eslint/parser": "^5.18.0",
+    "@angular-architects/module-federation": "^15.0.3",
+    "@angular-devkit/build-angular": "^15.1.2",
+    "@angular-eslint/eslint-plugin": "^15.2.0",
+    "@angular-eslint/eslint-plugin-template": "^15.2.0",
+    "@angular-eslint/template-parser": "^15.2.0",
+    "@angular/cli": "~15.1.0",
+    "@angular/compiler-cli": "^15.1.1",
+    "@angular/language-service": "^15.1.1",
+    "@nrwl/angular": "15.5.2",
+    "@nrwl/cli": "15.5.2",
+    "@nrwl/cypress": "15.5.2",
+    "@nrwl/eslint-plugin-nx": "15.5.2",
+    "@nrwl/jest": "15.5.2",
+    "@nrwl/linter": "15.5.2",
+    "@nrwl/workspace": "15.5.2",
+    "@types/jest": "^28.1.8",
+    "@types/node": "^18.11.18",
+    "@typescript-eslint/eslint-plugin": "^5.48.2",
+    "@typescript-eslint/parser": "^5.48.2",
     "codelyzer": "^6.0.2",
-    "cypress": "^9.5.4",
-    "dotenv": "^16.0.0",
-    "eslint": "^8.13.0",
-    "eslint-config-prettier": "^8.5.0",
+    "cypress": "^12.2.0",
+    "dotenv": "^16.0.3",
+    "eslint": "^8.32.0",
+    "eslint-config-prettier": "^8.6.0",
     "eslint-plugin-cypress": "^2.12.1",
-    "fake-indexeddb": "^3.1.7",
-    "jest": "27.5.1",
-    "jest-preset-angular": "^11.1.1",
-    "ng-packagr": "^13.3.0",
-    "nx": "14.0.3",
-    "postcss": "^8.4.12",
-    "postcss-import": "^14.1.0",
-    "postcss-preset-env": "^7.4.3",
+    "fake-indexeddb": "^4.0.1",
+    "jest": "^28.1.3",
+    "jest-environment-jsdom": "^28.1.3",
+    "jest-preset-angular": "^12.2.5",
+    "ng-packagr": "^15.1.1",
+    "nx": "15.5.2",
+    "postcss": "^8.4.21",
+    "postcss-import": "^15.1.0",
+    "postcss-preset-env": "^7.8.3",
     "postcss-url": "^10.1.3",
-    "prettier": "^2.6.2",
-    "ts-jest": "27.1.4",
-    "ts-node": "^10.7.0",
-    "typescript": "^4.6.3"
+    "prettier": "^2.8.3",
+    "ts-jest": "^28.0.8",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.4"
   },
   "optionalDependencies": {
     "fsevents": "^2.3.2"
