@@ -113,15 +113,8 @@ export class RemotePluginLoaderService {
   async loadModule(
     moduleDef: RemotePluginModuleOptions
   ): Promise<PluginModuleInterface> {
-    const module = await loadRemoteModule(moduleDef);
-    //console.log('Loaded Module:', module);
-/*    const mainModuleClass = module[moduleDef.moduleId];
-    if (mainModuleClass==null)  // The main module class is not defined
-    {
-      throw new Error("ModuleClass "+moduleDef.moduleId+" not exported in "+moduleDef.remoteEntry);
-    }*/
+    await loadRemoteModule(moduleDef);
     const mainModule = await this.compLoader.getOrCreatePluginModuleRef(moduleDef.moduleId.toLowerCase());
-    //createNgModuleRef<PluginModuleInterface>(mainModuleClass, this.injector).instance;
     return mainModule.instance;
   }
 
@@ -140,9 +133,7 @@ export class RemotePluginLoaderService {
       } catch (error) {
         console.error(
           'Error loading plugin ' +
-            moduleDef.moduleId +
-            ' from ' +
-            moduleDef.remoteEntry,
+            moduleDef.moduleId,
           error
         );
       }
